### 2020-05-22 #2192 @vsoch
Thank you @whedon, @arfon, and @usethedata! Since this is a Python-based tool that is generally usable across domains, I'll suggest reviewers that are keen on Python. I'll also try to choose reviewers that aren't currently reviewing, and haven't reviewed recently to help balance the work.

 - khinsen
 - ahurriyetoglu
 - Adrianzo
 - jochym
 - cbrueffer

Let me know if you want others, or otherwise different.
