### 2020-05-25 #2192 @usethedata
@vsoch -- Thank you.

@khninsen and @cbrueffer -- Would you be willing to review this paper for the Journal of Open Source Software?  I realize this would be a first review and I'm happy to help walk through that process.  When I have at least one (and preferably the needed two) reviewers, I will open up a review issue that will have a checklist for each reviewer and some additional instructions.  My email is in my GitHub profile if you'd like to discuss off-line.
