### 2020-05-31 #2192 @usethedata
@khinsen 👋   Pinging on this issue.  Are you able to review this paper for JOSS?
