### 2020-05-31 78e0911
adding pyflakes fixes and also as check for GitHub main workflow to ensure no unused imports!
Signed-off-by: vsoch <vsochat@stanford.edu>
