+++ gridtest/client/__init__.py
@@ -13,7 +13,6 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 import gridtest
 import argparse
 import sys
-import os
 
 
 def get_parser():
