+++ gridtest/logger/__init__.py
@@ -1,3 +1,7 @@
 from .message import bot
 from .progress import ProgressBar
 from .namer import RobotNamer
+
+assert bot
+assert ProgressBar
+assert RobotNamer
