+++ gridtest/client/check.py
@@ -8,8 +8,6 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 """
 
-import sys
-import os
 
 from gridtest.main.check import check_tests
 
