+++ gridtest/client/generate.py
@@ -9,7 +9,6 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 """
 
 import sys
-import os
 
 from gridtest.main.generate import generate_tests
 
