+++ .github/workflows/main.yml
@@ -14,13 +14,19 @@ jobs:
       - uses: actions/checkout@v2
 
       - name: Setup black environment
-        run: conda create --quiet --name black black
+        run: conda create --quiet --name black black pyflakes
 
       - name: Check formatting
         run: |
           export PATH="/usr/share/miniconda/bin:$PATH"
           source activate black
           black --check gridtest tests/*.py
+
+      - name: Check Imports
+        run: |
+          export PATH="/usr/share/miniconda/bin:$PATH"
+          source activate black
+          pyflakes gridtest
 
   testing:
     needs: formatting
