### 2020-05-31 vsoch/gridtest#26 @vsoch
All set! Please see https://github.com/vsoch/gridtest/pull/27. I also added pyflkes to run for the GitHub workflow so we won't miss future unused imports. And thanks for showing me this - I think I'm going to add it to most of my other python projects!

What we can do, if you agree, is to make changes for the review and merge into master, and then do a new release when the review is finished. So for the PR I've bumped the version, but won't actually release until after we are finished up.
