### 2020-05-31 vsoch/gridtest#26 @cbrueffer
I think that's a good solution, thanks for the quick fixes!
