### 2020-05-31 bdad604
adding pyflakes fixes and also as check for GitHub main workflow to ensure no unused imports! (#27)
Signed-off-by: vsoch <vsochat@stanford.edu>
