+++ gridtest/client/test.py
@@ -9,9 +9,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 """
 
 from gridtest.main.test import GridRunner
-import json
 import sys
-import re
 
 
 def main(args, extra):
