### 2020-06-01 #2192 @khinsen
@usethedata Yes, I can review this submission. Sorry for not reacting earlier, but GitHub notified me only today of being mentioned in this thread.
