### 2020-06-05 vsoch/gridtest#28 @khinsen
Clarification request: Who is this software for?
Background: I am looking at gridtest as a reviewer for JOSS (see [here](https://github.com/openjournals/joss-reviews/issues/2284)). A question I have been asking myself in this context is: what is the developer or user profile this package is aimed at? I haven't found the answer yet, and the section "Who is this software for?" in the README doesn't answer it either.

gridtest seems to be
 1. a lightweight testing framework for Python developers, with a focus on parameter scans
 2. a command-line tool to generate input parameter scans as JSON files, for use in testing or evaluation software that is not necessarily written in Python
 3. a Python library for generating grids of input parameters

The documentation is clearly written for Python programmers, which seems to exclude scenario 2, although other parts of the documentation refer to it. But the main problem I see is that the documentation mixes the three scenarios to the point that for most potential users, the first impression is "not sure this is really for me". And for me as a reviewer, that means I don't really know into whose shoes I am supposed to put myself.
