### 2020-06-05 vsoch/gridtest#28 @vsoch
It’s a very generic library, so the user group can generally be anyone that wants to package reproducible grids with their software (a research scientist) to someone that wants to use gridtest for testing (an RSE or research software engineer!)
