### 2020-06-08 vsoch/gridtest#28 @khinsen
Suggestion: remove the section "Who is this software for?" from the README, if the only possible answer is "it's for the people who have a use for this software".
