### 2020-06-08 9fc85f3
adding instructions to run tests in install documentation to close #29
Signed-off-by: vsoch <vsochat@stanford.edu>
