### 2020-06-08 vsoch/gridtest#28 @vsoch
I think it's good to have it clearly stated that it's not foremost a testing library, and I think that the section is useful to have, upfront. Could we perhaps work together to come up with a a more specific paragraph than "It's for the people who have a user for the software?" Here is a starting suggestion:

> Gridtest is intended for definition and saving of grids for any need that you might have,
or even for quick generation of running tests. It is not intended
to be a robust testing library like pytest or even unittest, but rather a quick
way to generate reproducible grids for scientific analyses.
