### 2020-06-09 vsoch/gridtest#28 @khinsen
I'd be happy to help with this, but I really have no idea of which user profile(s) gridtest targets. That's why I opened this issue!

Consider the "testing" use case. Most Python developers would go for the built-in `unittest`, and then perhaps complement it with `pytest` or `nose`. If they consider this framework too heavy, they'd write testing scripts of their own. So why should they use `gridtest` instead? It has some  advantages, of course, such as tests-as-yaml and of course support for grids. That has to weighed against the cost of "one more dependency" (compared to home-grown scripts) or of "different from mainstream" (compared to `unittest` and friends). What are the uses cases where `gridtest` represents the sweet spot? That's what needs to be in the README, because few people will read through the documentation to discover potential advantages of `gridtest` themselves.

BTW, speaking of grids, they are mentioned as a main selling argument, but no detail is given. When I first read "grid", I thought of `numpy.arange(...)` along multiple dimensions. I had to delve deep into the documentation to discover that `gridtest`'s concept of a grid is more flexible, and I still don't know how flexible it really is, even after having read most of the documentation.
