### 2020-06-09 vsoch/gridtest#28 @vsoch
I think perhaps a better solution would be to remove the extra docs from the README and direct the user to the rendered pages? The README I kept because I thought it was helpful but it seems to have had a negative experience for you. What do you think!
