### 2020-06-09 vsoch/gridtest#28 @vsoch
Personally, I wouldn't jump to use it in a robust testing sense for a complex software library - I also really like pytest for that. If I was still a research scientist and I had a small script I might use it for testing, however, because it would help me to quickly generate some basic tests. Generally though I'm attracted to the reproducibility side of things, so I think of gridtest as a generator tool that would let me package reproducible grids that might also allow for measurement of metrics. For example, I might create a set of grids to be used to generate inputs for machine learning, and then anyone can edit a particular function that trains / tests a model, and then gridtest can collect metrics like time to train, sensitivity and specificity, etc. I can see having repos of just grids for people to grab and plug their functions into different generator contexts. Yes, there are definitely other ways to do this, as there are many ways to skin a cat, but this is a different approach that starts to get people thinking about separating the parameter grids or metrics collection from some primary set of functions.

I'm not claiming this to be earth shattering software, but I think there is place for it in the research software universe, even if a much smaller niche than some other python library. As a developer the generality of the tool is exactly what excites me, because I'm wanting to wait and see what creative use cases people come up with! That includes myself - I do so much with testing and yaml and definitely see this being useful down the line.
