### 2020-06-09 #2284 @usethedata
@khinsen Thanks for your comments.  It is interesting to me that we've both read the documentation and code and come to some very different conclusions.  I confess that I struggle to understand the perspectives of "no clear statement need" and "no high level overview".  That's part of what's taken me a while to respond to your questions, since my read of the paper and documentation gave me a clear understanding of the problem(s) the author is working to solve and the overview of that solution.  There is also, to me, a very clear statement of functionality, which is the generation of grids, for which a primary purpose is testing software.

I realize that trying to explain things in this context (purely written) can be challenging, since all of the non-verbal clues to express context are missing.  But I do need to ask for more information about what you see is missing from the documentation noted above.

I understand the points you raise regarding reusable code versus re-editable code.  My thought is that, as you note, there is not a one size fits all answer to that balance.  A user is free to use this code either as a module (reusable, as presently developed) or from a copy and paste (re-editable).

This particular format (written only, lacking all of the non-verbals) can be challenging to convey complex subjects.  I would be happy to get on a video call with you (Zoom, GoToMeeting, Teams, WebEx, ...) to discuss, if you feel that would be useful.  I'm in US Eastern Time (UTC-04:00) and can adjust my schedule to find a time at your convenience.
