### 2020-06-09 vsoch/gridtest#28 @khinsen
@vsoch Lots of good ideas there! A README is always good to have, especially on GitHub where it figures so prominently. What I expect to find there is 1) a description of the project that lets me quickly see if it's worth (for me!) to have a closer look and 2) information on getting started, either in-line or as pointers to the main documentation. My first impression from the current README is "this is a piece of software that does many exciting things, but you have to figure them out for yourself".

I'd summarize your last comment as two use case scenarios:
 - low overhead tests for Python scripts and small packages
 - generating input data for reproducible computations

I hope I got that right - but it certainly is the kind of information I was looking for!
