### 2020-06-09 vsoch/gridtest#28 @vsoch
All set - I've updated the README and have a PR for your review: https://github.com/vsoch/gridtest/pull/30
