### 2020-06-18 vsoch/gridtest#31 @cbrueffer
Contributing guidelines
### Description

Tthe `Community guidelines` section of the JOSS review mentions instructions for contributors to the software. Adding a contributing file as per https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors would cover this.
