### 2020-06-18 vsoch/gridtest#31 @vsoch
I'd be happy to! Also note that we already have a guide in the documentation https://vsoch.github.io/gridtest/contributing/index.html. Would you prefer a standard CONTRIBUTING.md added, or perhaps a link to that page in the README?
