### 2020-06-18 vsoch/gridtest#31 @cbrueffer
To me, specific sections in the README are usually best, since they are very visible. I think the information in the guide is fine, so no CONTRIBUTING file is needed.
