### 2020-06-18 002de28
adding CONTRIBUTING.md and link to contributing docs in readme
Signed-off-by: vsoch <vsochat@stanford.edu>
