+++ .github/CONTRIBUTING.md
@@ -1,0 +1,111 @@
+# Contributor's Agreement
+
+This code is licensed under the MPL 2.0 [LICENSE](LICENSE). This contributing
+guide provides detailed instructions and a checklist for pull requests.
+Please also note contributing documentation available [here](https://vsoch.github.io/gridtest/contributing/index.html).
+
+# Contributing
+
+When contributing to GridTest, it is important to properly communicate the
+gist of the contribution. If it is a simple code or editorial fix, simply
+explaining this within the GitHub Pull Request (PR) will suffice. But if this
+is a larger fix or enhancement, it should be first discussed with the project
+team.
+
+Please note we have a code of conduct, described below. Please follow it in
+all your interactions with the project members and users.
+
+## Pull Request Process
+
+1. All PRs should be against the master branch.
+2. Follow the existing code style precedent. We use black for linting, and you
+   can install it `pip install black` and then run `black gridtest` to format the code.
+   This will be tested.
+3. Test your PR locally, and provide the steps necessary to test for the
+   reviewers.
+4. The project's default copyright and header have been included in any new
+   source files.
+5. All (major) changes to GridTest must be documented in the docs folder.
+   If your PR changes a core functionality, please 
+   include clear description of the changes in your PR so that the docs 
+   can be updated, or better, submit another PR to update the docs directly.
+6. Update the CHANGELOG.md. If necessary, update the README.md.
+7. The pull request will be reviewed by others, and the final merge must be
+   done by the project lead, @vsoch (or approved by her).
+
+
+# Code of Conduct
+
+## Our Pledge
+
+In the interest of fostering an open and welcoming environment, we as
+contributors and maintainers pledge to making participation in our project and
+our community a harassment-free experience for everyone, regardless of age, body
+size, disability, ethnicity, gender identity and expression, level of experience,
+nationality, personal appearance, race, religion, or sexual identity and
+orientation.
+
+## Our Standards
+
+Examples of behavior that contributes to creating a positive environment
+include:
+
+* Using welcoming and inclusive language
+* Being respectful of differing viewpoints and experiences
+* Gracefully accepting constructive criticism
+* Focusing on what is best for the community
+* Showing empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+* The use of sexualized language or imagery and unwelcome sexual attention or
+  advances
+* Trolling, insulting/derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others' private information, such as a physical or electronic
+  address, without explicit permission
+* Other conduct which could reasonably be considered inappropriate in a
+  professional setting
+
+### Our Responsibilities
+
+Project maintainers are responsible for clarifying the standards of acceptable
+behavior and are expected to take appropriate and fair corrective action in
+response to any instances of unacceptable behavior.
+
+Project maintainers have the right and responsibility to remove, edit, or
+reject comments, commits, code, wiki edits, issues, and other contributions
+that are not aligned to this Code of Conduct, or to ban temporarily or
+permanently any contributor for other behaviors that they deem inappropriate,
+threatening, offensive, or harmful.
+
+## Scope
+
+This Code of Conduct applies both within project spaces and in public spaces
+when an individual is representing the project or its community. Examples of
+representing a project or community include using an official project e-mail
+address, posting via an official social media account, or acting as an appointed
+representative at an online or offline event. Representation of a project may be
+further defined and clarified by project maintainers.
+
+## Enforcement
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be
+reported by contacting the project leader @vsoch. All
+complaints will be reviewed and investigated and will result in a response
[…]
