### 2020-06-18 vsoch/gridtest#31 @vsoch
All set! https://github.com/vsoch/gridtest/commit/002de28e6529d99082baab9864027974c3c361a5 I did the following:

 - added a CONTRIBUTING.md guide which has more of a "PR checklist" along with code of conduct
 - linked to the "how to contribute to docs and code" in the README, along with the link to the checklist.

I figure there are two use cases to wanting to contribute - either wanting the checklist (CONTRIBUTING.md) or the pretty, rendered guide, so we can link to both.
