### 2020-06-18 vsoch/gridtest#31 @cbrueffer
Looks good, thanks! I think moving the pointer to the documentation from the end of the overview to its own Documentation section would be good to make it more obvious. It gets a bit lost in its current location.
