### 2020-06-18 7d97d88
adding documentation link to top of repo
Signed-off-by: vsoch <vsochat@stanford.edu>
