### 2020-06-18 38dad3a
adding documentation link to top of readme
Signed-off-by: vsoch <vsochat@stanford.edu>
