### 2020-06-18 vsoch/gridtest#31 @cbrueffer
Didn't even notice that! It's probably a very personal thing on what we focus on first when skimming a repo.
