### 2020-06-18 vsoch/gridtest#31 @cbrueffer
Off-topic for this PR but I noticed that gridtest isn't tested using gridtest from the CI, is there a reason for that?
