### 2020-06-20 c100da5
Various API documentation improvements.
Various API documentation improvements, including correcting arguments,
and typo fixes.
