### 2020-06-20 vsoch/gridtest#31 @cbrueffer
Addressed in https://github.com/vsoch/gridtest/commit/002de28e6529d99082baab9864027974c3c361a5, https://github.com/vsoch/gridtest/commit/7d97d88fdefe6554158b403db22ef1dc31b0ab79, and https://github.com/vsoch/gridtest/commit/38dad3a2f5eeac70178158c258b5e874c83b8621
