### 2020-06-20 ba836c0
Fix some spelling, grammar, and punctuation in the paper.
