### 2020-06-20 0c1907a
Various API documentation improvements. (#32)
Various API documentation improvements, including correcting arguments,
and typo fixes.
