### 2020-06-20 1b13311
Fix some spelling, grammar, and punctuation in the paper. (#33)
* Fix some spelling, grammar, and punctuation in the paper.
* Revert quotation mark change.
