### 2020-06-20 vsoch/gridtest#34 @cbrueffer
Paper: discussion of state of the art
### Description

Reading the paper with the JOSS review checklist in mind, one thing I find missing is a discussion of the current state of the art, both in terms GridTest as a testing tool, and as a grid generation tool.

From an online search, there seem to be several software tools in the general "grid generation" space and even some theory behind it. Examples would be https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ParameterGrid.html and (non-Python) https://rdrr.io/cran/dials/man/grid_regular.html.
