### 2020-06-20 vsoch/gridtest#34 @vsoch
I can definitely add a few lines to compare these! Give me a few minutes.
