### 2020-06-20 vsoch/gridtest#34 @cbrueffer
The two mentioned above are just examples from a quick search; what would be nice is a bit more thorough survey of what's out there, how GridTest compares to them, and what it adds. I think this goes a bit in the direction of what @khinsen referred to to with "statement of need".
