### 2020-06-20 3e92e30
adding references to other libraries for grids
Signed-off-by: vsoch <vsochat@stanford.edu>
