+++ paper/paper.bib
@@ -22,3 +22,63 @@
   booktitle={SQAMIA},
   year={2015}
 }
+
+@MISC{sklearn,
+  title        = "{sklearn.model\_selection.ParameterGrid} --- scikit-learn
+                  0.23.1 documentation",
+  abstract     = "scikit-learn: machine learning in Python",
+  howpublished = "\url{https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ParameterGrid.html}",
+  note         = "Accessed: 2020-6-20"
+}
+
+@MISC{grid_regular,
+  title        = "grid\_regular: Create grids of tuning parameters in dials:
+                  Tools for Creating Tuning Parameter Values",
+  author       = "[aut, Max Kuhn and {cre]} and [cph], Rstudio",
+  abstract     = "Random and regular grids can be created for any number of
+                  parameter objects.",
+  month        =  apr,
+  year         =  2020,
+  howpublished = "\url{https://rdrr.io/cran/dials/man/grid_regular.html}",
+  note         = "Accessed: 2020-6-20"
+}
+
+
+@MISC{keras,
+  title        = "How to Grid Search Hyperparameters for Deep Learning Models
+                  in Python With Keras - Machine Learning Mastery",
+  booktitle    = "Machine Learning Mastery",
+  author       = "Brownlee, Jason",
+  abstract     = "Hyperparameter optimization is a big part of deep learning.
+                  The reason is that neural networks are notoriously difficult
+                  to configure and there are a lot of parameters that need to
+                  be set. On top of that, individual models can be very slow to
+                  train. In this post you will discover how you can use the
+                  grid",
+  month        =  aug,
+  year         =  2016,
+  howpublished = "\url{https://machinelearningmastery.com/grid-search-hyperparameters-deep-learning-models-python-keras/}",
+  note         = "Accessed: 2020-6-20"
+}
+
+
+@MISC{h2o,
+  title        = "Grid (Hyperparameter) Search --- {H2O} 3.30.0.5 documentation",
+  howpublished = "\url{https://docs.h2o.ai/h2o/latest-stable/h2o-docs/grid-search.html}",
+  note         = "Accessed: 2020-6-20",
+  year         = 2020
+}
+
+
+@MISC{Willkoehrsen2018-hm,
+  title        = "Intro to Model Tuning: Grid and Random Search",
+  booktitle    = "Kaggle",
+  author       = "{willkoehrsen}",
+  abstract     = "Explore and run machine learning code with Kaggle Notebooks |
+                  Using data from multiple data sources",
+  month        =  jul,
+  year         =  2018,
+  howpublished = "\url{https://kaggle.com/willkoehrsen/intro-to-model-tuning-grid-and-random-search}",
+  note         = "Accessed: 2020-6-20"
+}
+
