### 2020-06-20 #2284 @vsoch
Thank you @cbrueffer! It's always really a wonderful surprise when a reviewer takes the initiative to help out with changes, and I really appreciate that. The last issue (for https://github.com/vsoch/gridtest/issues/34) now has a PR open to address it: https://github.com/vsoch/gridtest/pull/35.
