### 2020-06-22 d7767cf
add getting started and index page achorlinks
Signed-off-by: vsoch <vsochat@stanford.edu>
