### 2020-06-22 2e33bdc
add getting started and index page achorlinks
Signed-off-by: vsoch <vsochat@stanford.edu>
