### 2020-06-22 #2284 @vsoch
Sounds good! Here is a PR with the suggested changes https://github.com/vsoch/gridtest/pull/36.
