### 2020-06-23 df038ca
updating grids definition (#36)
Signed-off-by: vsoch <vsochat@stanford.edu>
