### 2020-06-29 50a3b1f
fixing citations
Signed-off-by: vsoch <vsochat@stanford.edu>
