+++ paper/paper.bib
@@ -23,43 +23,40 @@
   year={2015}
 }
 
-@MISC{sklearn,
-  title        = "{sklearn.model\_selection.ParameterGrid} --- scikit-learn
-                  0.23.1 documentation",
-  abstract     = "scikit-learn: machine learning in Python",
-  howpublished = "\url{https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ParameterGrid.html}",
-  note         = "Accessed: 2020-6-20"
+@article{sklearn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
 }
 
-@MISC{grid_regular,
-  title        = "grid\_regular: Create grids of tuning parameters in dials:
-                  Tools for Creating Tuning Parameter Values",
-  author       = "[aut, Max Kuhn and {cre]} and [cph], Rstudio",
-  abstract     = "Random and regular grids can be created for any number of
-                  parameter objects.",
-  month        =  apr,
-  year         =  2020,
-  howpublished = "\url{https://rdrr.io/cran/dials/man/grid_regular.html}",
-  note         = "Accessed: 2020-6-20"
+@article{grid_regular,
+author = "Beilsten-Edmands, James and Winter, Graeme and Gildea, Richard and Parkhurst, James and Waterman, David and Evans, Gwyndaf",
+title = "{Scaling diffraction data in the {\it DIALS} software package: algorithms and new approaches for multi-crystal scaling}",
+journal = "Acta Crystallographica Section D",
+year = "2020",
+volume = "76",
+number = "4",
+pages = "385--399",
+month = "Apr",
+doi = {10.1107/S2059798320003198},
+url = {https://doi.org/10.1107/S2059798320003198},
+abstract = {In processing X-ray diffraction data, the intensities obtained from integration of the diffraction images must be corrected for experimental effects in order to place all intensities on a common scale both within and between data collections. Scaling corrects for effects such as changes in sample illumination, absorption and, to some extent, global radiation damage that cause the measured intensities of symmetry-equivalent observations to differ throughout a data set. This necessarily requires a prior evaluation of the point-group symmetry of the crystal. This paper describes and evaluates the scaling algorithms implemented within the {\it DIALS} data-processing package and demonstrates the effectiveness and key features of the implementation on example macromolecular crystallographic rotation data. In particular, the scaling algorithms enable new workflows for the scaling of multi-crystal or multi-sweep data sets, providing the analysis required to support current trends towards collecting data from ever-smaller samples. In addition, the implementation of a free-set validation method is discussed, which allows the quantification of the suitability of scaling-model and algorithm choices.},
+keywords = {diffraction, crystallography, multi-crystal, data analysis, scaling},
 }
 
-
-@MISC{keras,
-  title        = "How to Grid Search Hyperparameters for Deep Learning Models
-                  in Python With Keras - Machine Learning Mastery",
-  booktitle    = "Machine Learning Mastery",
-  author       = "Brownlee, Jason",
-  abstract     = "Hyperparameter optimization is a big part of deep learning.
-                  The reason is that neural networks are notoriously difficult
-                  to configure and there are a lot of parameters that need to
[…]
