### 2020-06-29 #2284 @vsoch
@khinsen I'll get on this! I was slow to respond to https://github.com/vsoch/gridtest/pull/35 (I didn't see the comment!) so once that is merged, I'll also work on the use cases section.
