### 2020-06-30 50076c3
adding references to other libraries for grids (#35)
* adding references to other libraries for grids

Signed-off-by: vsoch <vsochat@stanford.edu>

* fixing citations

Signed-off-by: vsoch <vsochat@stanford.edu>
