+++ paper/paper.bib
@@ -22,3 +22,60 @@
   booktitle={SQAMIA},
   year={2015}
 }
+
+@article{sklearn,
+ title={Scikit-learn: Machine Learning in {P}ython},
+ author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ journal={Journal of Machine Learning Research},
+ volume={12},
+ pages={2825--2830},
+ year={2011}
+}
+
+@article{grid_regular,
+author = "Beilsten-Edmands, James and Winter, Graeme and Gildea, Richard and Parkhurst, James and Waterman, David and Evans, Gwyndaf",
+title = "{Scaling diffraction data in the {\it DIALS} software package: algorithms and new approaches for multi-crystal scaling}",
+journal = "Acta Crystallographica Section D",
+year = "2020",
+volume = "76",
+number = "4",
+pages = "385--399",
+month = "Apr",
+doi = {10.1107/S2059798320003198},
+url = {https://doi.org/10.1107/S2059798320003198},
+abstract = {In processing X-ray diffraction data, the intensities obtained from integration of the diffraction images must be corrected for experimental effects in order to place all intensities on a common scale both within and between data collections. Scaling corrects for effects such as changes in sample illumination, absorption and, to some extent, global radiation damage that cause the measured intensities of symmetry-equivalent observations to differ throughout a data set. This necessarily requires a prior evaluation of the point-group symmetry of the crystal. This paper describes and evaluates the scaling algorithms implemented within the {\it DIALS} data-processing package and demonstrates the effectiveness and key features of the implementation on example macromolecular crystallographic rotation data. In particular, the scaling algorithms enable new workflows for the scaling of multi-crystal or multi-sweep data sets, providing the analysis required to support current trends towards collecting data from ever-smaller samples. In addition, the implementation of a free-set validation method is discussed, which allows the quantification of the suitability of scaling-model and algorithm choices.},
+keywords = {diffraction, crystallography, multi-crystal, data analysis, scaling},
+}
+
+@online{keras,
+  title={Keras},
+  author={Chollet, Francois and others},
+  year={2015},
+  publisher={GitHub},
+  url={https://github.com/fchollet/keras},
+} 
+
+
+@MISC{h2o,
+  title        = "Grid (Hyperparameter) Search --- {H2O} 3.30.0.5 documentation",
+  howpublished = "\url{https://docs.h2o.ai/h2o/latest-stable/h2o-docs/grid-search.html}",
+  note         = "Accessed: 2020-6-20",
+  year         = 2020
+}
+
+
+@MISC{Willkoehrsen2018-hm,
+  title        = "Intro to Model Tuning: Grid and Random Search",
+  booktitle    = "Kaggle",
+  author       = "{willkoehrsen}",
+  abstract     = "Explore and run machine learning code with Kaggle Notebooks |
+                  Using data from multiple data sources",
+  month        =  jul,
+  year         =  2018,
+  howpublished = "\url{https://kaggle.com/willkoehrsen/intro-to-model-tuning-grid-and-random-search}",
+  note         = "Accessed: 2020-6-20"
+}
+
