### 2020-06-30 91ac412
updating use cases in paper
Signed-off-by: vsoch <vsochat@stanford.edu>
