+++ paper/paper.bib
@@ -23,6 +23,17 @@
   year={2015}
 }
 
+
+@MISC{sklearn-tutorial,
+  title        = "{sklearn.model\_selection.ParameterGrid} --- scikit-learn
+                  0.23.1 documentation",
+  abstract     = "scikit-learn: machine learning in Python",
+  howpublished = "\url{https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ParameterGrid.html}",
+  note         = "Accessed: 2020-6-30",
+  year="{2020}"
+}
+
+
 @article{sklearn,
  title={Scikit-learn: Machine Learning in {P}ython},
  author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
