### 2020-06-30 76cca5a
use semicolons
Signed-off-by: vsoch <vsochat@stanford.edu>
