+++ paper/paper.bib
@@ -46,7 +46,7 @@
  year={2011}
 }
 
-@article{grid_regular,
+@article{gridregular,
 author = "Beilsten-Edmands, James and Winter, Graeme and Gildea, Richard and Parkhurst, James and Waterman, David and Evans, Gwyndaf",
 title = "{Scaling diffraction data in the {\it DIALS} software package: algorithms and new approaches for multi-crystal scaling}",
 journal = "Acta Crystallographica Section D",
