### 2020-06-30 2c274b4
add link to concept section
Signed-off-by: vsoch <vsochat@stanford.edu>
