### 2020-06-30 972cc4a
move use cases down under concepts
Signed-off-by: vsoch <vsochat@stanford.edu>
