+++ paper/paper.bib
@@ -30,7 +30,7 @@
   abstract     = "scikit-learn: machine learning in Python",
   howpublished = "\url{https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ParameterGrid.html}",
   note         = "Accessed: 2020-6-30",
-  year="{2020}"
+  year         = 2020
 }
 
 
