### 2020-06-30 #2284 @vsoch
@khinsen please take a look! I've done the following changes:
 - extended the background section to lead into the idea of "grids as first class citizens"
 - moved the concepts section to appear _before_ use cases
 - extended the use cases to include the (somewhat modified) content that we added to the readme, etc. about the different use cases for grids.

Please take a look and let me know your feedback!
