### 2020-07-02 #2284 @khinsen
@vsoch Thanks, that looks good! I like the many links to further details on the Web site.
