### 2020-07-02 ed4284f
Minor suggestions for improving the text
Feel free to cherry-pick!
