### 2020-07-02 d0e7b31
Update paper/paper.md
