### 2020-07-02 7f7860a
Minor suggestions for improving the text (#37)
* Minor suggestions for improving the text
Feel free to cherry-pick!
* Update paper/paper.md
Co-authored-by: Vanessasaurus <vsochat@gmail.com>
