### 2020-07-02 a30821a
Fix itemization; capitalize acronyms.
