### 2020-07-02 09778a9
Add a missing word and small fixes for the references.
