+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @MISC{gridtest,
-  title        = "GridTest on GitHub",
+  title        = "{GridTest on GitHub}",
   booktitle    = "GridTest",
   abstract     = "https://github.com/vsoch/gridtest",
   year         =  2020,
@@ -81,7 +81,7 @@ keywords = {diffraction, crystallography, multi-crystal, data analysis, scaling}
 @MISC{Willkoehrsen2018-hm,
   title        = "Intro to Model Tuning: Grid and Random Search",
   booktitle    = "Kaggle",
-  author       = "{willkoehrsen}",
+  author       = "{Koehrsen, Will}",
   abstract     = "Explore and run machine learning code with Kaggle Notebooks |
                   Using data from multiple data sources",
   month        =  jul,
