### 2020-07-02 07368af
Various small fixes for the paper (#38)
* Fix itemization; capitalize acronyms.
* Add a missing word and small fixes for the references.
