### 2020-07-02 #2284 @vsoch
New changes from @cbrueffer !
