### 2020-07-03 f5fcfa5
Another try to get the citation style consistent. (#39)
