+++ paper/paper.bib
@@ -81,7 +81,7 @@ keywords = {diffraction, crystallography, multi-crystal, data analysis, scaling}
 @MISC{Willkoehrsen2018-hm,
   title        = "Intro to Model Tuning: Grid and Random Search",
   booktitle    = "Kaggle",
-  author       = "{Koehrsen, Will}",
+  author       = {Koehrsen, W.},
   abstract     = "Explore and run machine learning code with Kaggle Notebooks |
                   Using data from multiple data sources",
   month        =  jul,
