### 2020-07-04 vsoch/gridtest#40 @usethedata
Errors/suggestions needed for JOSS publication of paper.
### Description

This issue lists corrections and suggestions for items related to the publication of the Gridtest paper in JOSS.

### Corrections requested

- [x] In the paper, the link for "how does it work" is 404 (https://vsoch.github.io/gridtest/getting-started/index.html%23how-does-it-work)
- [x] In the paper, the link for "loading grids via a GridRunner is 404 (https://vsoch.github.io/gridtest/getting-started/grids/index.html%23loading-via-a-gridrunner)

### Additional steps needed

- [x] Once the above corrections have been made and committed, make a tagged release.
- [x] Archive the tagged release, e.g. with [Figshare](https://figshare.com/) or [Zenodo](https://zenodo.org/)
- [x] Post the version tag and the DOI for the archived version on the [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/2284)

Once those items are done, I can move to my next steps.
