### 2020-07-04 #2284 @usethedata
@vsoch -- I've opened https://github.com/vsoch/gridtest/issues/40 to request some minor corrections to the paper and a list of the steps needed from you to move this forward.  Thanks much.
