### 2020-07-04 vsoch/gridtest#40 @vsoch
working on these now!
