### 2020-07-04 5fde29c
fixing links that 404
Signed-off-by: vsoch <vsochat@stanford.edu>
