### 2020-07-04 vsoch/gridtest#40 @vsoch
- [x] Links are [fixed](https://github.com/vsoch/gridtest/pull/41) (I checked the branch with the Joss Paper Preview / Whedon tool)
 - [x] A [new release is made](https://github.com/vsoch/gridtest/releases/tag/0.0.15) although the [doi seems to not be ready yet](https://doi.org/10.5281/zenodo.3930366). I needed to update my author name so I'll wait for that page to render to do that. I'll still post both on the JoSS issue, and please close here when you are happy with the changes!
