### 2020-07-04 #2284 @vsoch
- Release is created under zenodo at https://doi.org/10.5281/zenodo.3930366 (URL seems to not be working yet)
 - Release version is [0.0.15](https://github.com/vsoch/gridtest/releases/tag/0.0.15)

I needed to change my name in the record but it doesn't appear I can do that yet! I'm okay if it stays like that either way.
