### 2020-07-04 #2284 @vsoch
okay! the doi appeared and I was able to update name https://zenodo.org/record/3930366#.XwDI35ZME5k so I think we are good.
