### 2020-07-05 #2284 @danielskatz
@vsoch - I'm the AEiC on duty currently, and this mostly looks good to go.  One exception is the use of second person, which I would prefer to see as third person for a scientific article, as opposed to a tutorial or user documentation, where second person would be fine.  Would you be willing to change this?
