### 2020-07-05 b6e2eed
removing second person from paper
Signed-off-by: vsoch <vsochat@stanford.edu>
