+++ paper/paper.md
@@ -60,8 +60,8 @@ realized that she might want to define just sets of parameters that expand
 into matrices that can be useful across many use cases. This makes
 the grids "first class citizens." For example, instead of a top to bottom
 script that loops over some set of datasets, parameters, and algorithms, 
-you could define grids to generate each in a modular fashion. This
-is explained in detail for the [clustering grids](https://github.com/vsoch/gridtest/tree/master/examples/clustering-grids) example derived from scikit-learn. As another example, you might
+the user could define grids to generate each in a modular fashion. This
+is explained in detail for the [clustering grids](https://github.com/vsoch/gridtest/tree/master/examples/clustering-grids) example derived from scikit-learn. As another example, the user might
 just want to parameterize some set of inputs to randomly generate a cohort.
 This example is detailed [in another tutorial here](https://vsoch.github.io/gridtest/tutorials/samplegrid/). The overall idea is simple. The current practice is generally to write parameterizations alongside code, whether that means nested for loops or decorators for testing.
 GridTest allows for this same kind of functionality, but storing the parameterization
@@ -81,12 +81,11 @@ For an example of command line usage, the reader is directed to the ["How does i
 
 ### Parameterization
 
-A **grid** is a global definition of a parameter matrix. You can define arguments,
+A **grid** is a global definition of a parameter matrix. A user can define arguments,
 and optionally functions to run to be mapped to arguments. Grids are generated
-on demand, meaning when you iterate over a grid object, so that they are more
-efficient to use because we don't store a large list in memory.
-Grids can be put to many uses. You might share a repository that only defines grids that people
-can use across many different kinds of machine learning models, likely to collect metrics
+on demand, meaning when the user iterates over a Grid object, so that no large lists are stored in memory.
+Grids can be put to many uses. The user might share a repository that only defines grids that people
+can use across many different kinds of machine learning models, possibly to collect metrics
 and compare different analysis strategies being used. An introduction to grids
 is available [here](https://vsoch.github.io/gridtest/getting-started/grids/).
 
@@ -94,8 +93,8 @@ is available [here](https://vsoch.github.io/gridtest/getting-started/grids/).
 
 A **metric** is a Python decorator that is paired with a test that will measure some
 attribute of a test. For example:
- - you might run a function across a grid of arguments, and then measure the time that each combination takes (the metric), and generate a report for inspection.
- - you might be doing text processing and having functions to parse text. Each function might be run over a grid of sentences and counts, and for each result, we want to count the number of unique words, and total words (metrics).
+ - the user might run a function across a grid of arguments, and then measure the time that each combination takes (the metric), and generate a report for inspection.
+ - the user might be doing text processing and having functions to parse text. Each function might be run over a grid of sentences and counts, and for each result, the number of unique and total words is counted (metrics).
 
 Metrics are fully described in the [metrics](https://vsoch.github.io/gridtest/getting-started/metrics/) section
 of the documentation.
@@ -114,8 +113,8 @@ GridTest extends the traditional definition [@sklearn-tutorial] of a grid to inc
 
  - [generating random samples](https://vsoch.github.io/gridtest/tutorials/samplegrid/)
  - [loading grids via a GridRunner](https://vsoch.github.io/gridtest/getting-started/grids/index.html) class separate from the application's Python code.
- - generating grids as you go (meaning as an iterator)
- - previewing grids on the command line before you use them
+ - generating grids as they are needed (meaning as an iterator)
+ - previewing grids on the command line before using them
  - generating content of grids via external functions, and optionally unwrapping list values
 
[…]
