### 2020-07-05 ce4d3d1
tweak to paper
Signed-off-by: vsoch <vsochat@stanford.edu>
