### 2020-07-05 #2284 @vsoch
Fixed! https://github.com/vsoch/gridtest/commit/ce4d3d1868482559a833072f15f47e61a7d97514

That particular word has always been trouble for me - thanks for the catch!
