### 2020-07-05 #2284 @vsoch
Woohoo!! Thank you @usethedata @khinsen @cbrueffer @whedon and @danielskatz - this was a lot of fun!
